Move tcp functions into separate file
