+++ lib/transport-utils.js
@@ -5,6 +5,7 @@
 // Load modules
 
 var Util = require('util')
+var Stream = require('stream')
 var _ = require('lodash')
 var Nid = require('nid')
 var Patrun = require('patrun')
@@ -25,7 +26,8 @@ var internals = {
       'unknown_message_id': 'The message has an unknown identifier',
       'own_message': 'Inbound message rejected as originated from this server.',
       'message_loop': 'Inbound message rejected as looping back to this server.',
-      'data_error': 'Inbound message included an error description.'
+      'data_error': 'Inbound message included an error description.',
+      'invalid_json': 'Invalid JSON: <%=input%>.'
     },
     override: true
   })
@@ -527,6 +529,61 @@ internals.Utils.prepareResponse = function (seneca, input) {
     }
   }
 }
+
+internals.Utils.jsonStreamParser = function (seneca) {
+  var parser = new Stream.Duplex({ objectMode: true })
+  parser.linebuf = []
+  parser._read = function () {}
+  parser._write = function (data, enc, callback) {
+    var str = '' + data
+    var endline = -1
+    var remain = 0
+
+    while ((endline = str.indexOf('\n', remain)) !== -1) {
+      this.linebuf.push(str.substring(remain, endline))
+      var jsonstr = this.linebuf.join('')
+
+      this.linebuf.length = 0
+      remain = endline + 1
+
+      if (jsonstr === '') {
+        return callback()
+      }
+
+      var outdata = internals.Utils.parseJSON(seneca, 'stream', jsonstr)
+
+      if (outdata) {
+        this.push(outdata)
+      }
+    }
+
+    if (endline === -1) {
+      this.linebuf.push(str.substring(remain))
+    }
+
+    return callback()
+  }
+
+  return parser
+}
+
+internals.Utils.jsonStringifyStream = function (seneca) {
+  var stringify = new Stream.Duplex({ objectMode: true })
+  stringify._read = function () {}
+  stringify._write = function (data, enc, callback) {
+    var out = internals.Utils.stringifyJSON(seneca, 'stream', data)
+
+    if (out) {
+      this.push(out + '\n')
+    }
+
+    callback()
+  }
+
+  return stringify
+}
+
+internals.Utils.error = internals.error
 
 // Needed for transport exports of utils
 internals.Utils.prototype.close = internals.Utils.close
